requirements.txt
hatchet_sdk==0.38.3
numpy==2.1.3
torch==2.5.1
pymediainfo==6.1.0
transformers==4.46.3
pandas==2.2.3
boto3==1.35.73
pillow_heif==0.21.0
scipy==1.14.1
opencv-python==4.10.0.84
pyarrow==18.1.0
duckdb==1.1.3
torchvision==0.20.1
timm==1.0.12
einops==0.8.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pgvector==0.3.6
awscrt==0.22.0
# img2dataset==1.45.0
bs4==0.0.2
requests==2.32.3
onnxruntime==1.20.1
torchdata==0.10.1
s3torchconnector==1.3.2
vllm==0.7.2
